python/agents/auto-insurance-agent/pyproject.toml
[project]
name = "auto-insurance-agent"
version = "0.1.0"
description = "AI agent for auto insurance using Apigee API Hub for tools"
authors = [
    {name = "David Rush",email = "davidrush@google.com"}
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=0.5.0,<1.0.0)",
    "google-cloud-secret-manager (>=2.23.3,<3.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
google-adk = "^0.5.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.2"
google-cloud-aiplatform = { version = ">=1.94.0,<2.0.0", extras = ["adk","agent-engines"] }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
